refactor(l5): move maintenance-queue and maintenance-sweep from src/ into server/maintenance (issue 864)
L5 of the server-hardening ladder: cut two more modules out of src/ so
server/main.ts stops reaching back into it.

Moved (git mv, then split by responsibility to meet the server/ file rule):

- src/maintenance-queue.ts (1018 lines) ->
  - server/maintenance/maintenance-job.ts (job shape, error category,
    terminal-error marker, the queue port, row mapping, enqueue validation,
    backoff)
  - server/maintenance/maintenance-queue.ts (the MaintenanceQueue class and
    its SQL)
  - server/maintenance/maintenance-queue-runner.ts (MaintenanceQueueRunner,
    its logger and handler types, lease heartbeat)
- src/maintenance-sweep.ts (497 lines) ->
  server/maintenance/maintenance-sweep.ts

Both old src/ paths become L5 shims that re-export the server/ modules, so
every src/ and scripts/ importer keeps its import line untouched and stays
out of the pre-commit lint set.

server/ importers repointed: server/maintenance/index.ts,
server/application/index.ts, server/domain/candidate-dedupe.ts,
server/domain/candidate-dedupe.test.ts, server/maintenance/maintenance.pg.test.ts.

Lint work the server/ standard forced, no behavior change:

- validateEnqueue split into validateJobIdentity, validateRetryPolicy, and
  validateScope (complexity 22 -> under the rule); the checks and their
  error strings are identical and run in the same order.
- The distill lane-batch SQL hoisted to a module constant so
  selectDistillLaneBatches fits the function rule; the query text is byte
  identical.
- The one non-null assertion on a claim lease token replaced with a typed
  guard that throws on the impossible null instead of asserting it away.
- Promise.all no longer spreads an already-iterable Set; the self-referencing
  active promise is a const.

Tests left in src/ (retire or move at L6), all green through the shims:
src/maintenance-queue.test.ts, src/maintenance-queue.pg.test.ts,
src/maintenance-sweep.test.ts, src/maintenance-sweep.live.test.ts,
src/embedding-repair-formula.pg.test.ts,
src/db/migrations/026_maintenance_queue-retry.test.ts and its test helpers.

Runtime closure of src/ under server/main.ts: 46 -> 44.

Done-means: scripts/done-means/864-moved-out-of-src.sh src/maintenance-queue.ts src/maintenance-sweep.ts
